Add new windows path parser function: path_sep() (parser-iter)
During development of various windows path parser functions, a common
pattern has emerged where a parser that matches either the path
separator or the end of the byte string is needed. This is currently
hard coded into each parser function that needs this, however this
introduces a lot of redundant code.

This change adds a new parser function `path_sep` that, as mentioned,
matches either the path separator or the end of the byte string. With
the new function, future commits will then be able to remove hard-coding
this parser improving the maintainability of the code base.