Add input lifetime bound to windows parsers (parser-iter)
With the introduction of the path_sep windows parser function that
requires an explicit lifetime bound on its input, potentially every
other windows parser function will need the same bound as well due to
core parser functions such as nondevice_part() and component() using
many of the other parser functions. Without this bound, attempting to
use the path_sep() parser function will cause the compiler to complain
about the missing lifetime bound and refusing to build.

This change adds an explicit lifetime bound to the input of all relevant
windows parser functions. The functions where this was not added are
either going to have this lifetime bound added in a future commit -- and
this due only to massive changes to these functions eg nondevice_part(),
component().