fix: windows path parser reporting wrong position of errors
With the parser-iter feature active, when using iterating over a windows
path using the parser iterator, if a parsing error occurs, the position
of the error in the path that is recorded is incorrect. This is due to a
path component being interpreted through multiple parsers instead of
just a single parser. As a result, each individual parser only knows the
position of the end of the component instead of the exact position the
error occurred at.

For example, given an invalid path `/hello/world/hello:world`, the error
position is expected to be the value 19, however the recorded position
is instead 24 ie the end of the component that generated the error.

This change refactors many of the windows path parser functions to
ensure that only a single parser is being used to parse path components.
With this change, the correct position of a windows path parsing error
is stored and available to use when constructing an error message.

Please note that several parser functions are removed by this change as
they are not used at all anywhere in the codebase. The removed functions
are:

- valid_part_char()

- file_parts()

- file_name()