Add new ParseError method: msg()
When testing error objects such as ErrorInfo and ParseError, it is
sometimes needed to construct an error message manually using
information stored in the error objects and then comparing that to the
error message generated by the error objects. One part of this is the
string stored in the error objects that forms the base of the error
message. However, the string stored in the ParseError object cannot be
accessed, and thus testing of error messages when converting from an
ErrorInfo object to a ParseError object cannot be done.

This change adds a new public method to the ParseError struct `msg()`.
This new method simply returns a string slice of the owned string stored
in the ParseError object. This allows being able to test error messages
generated by the error objects.