Ignore benchmark code when collecting code coverage
Code coverage reports a lower percentage of code covered by tests than
expected. This is due to the addition of code used for benchmarking
performance of the library which is included in code coverage reports.

This change updates the justfile configuration so that benchmark code is
not included when calculating code coverage.