fix: unable to access crate::common module
It is intended for the common module and its submodules to be publically
accessible and usable. From the perspective of the project, this is
important for unit and integration tests, but this is also important for
users and/or contributors of the library who may want to create their
own path implementations. However, the common module is private and
inaccessible.

This change makes the common module publically accessible. The immediate
benefit of this is being able to more fully create unit and integration
tests.