fix: windows path storing wrong end position in error
When creating an ErrorInfo object to represent an error parsing a
windows path, the position stored representing the end position in the
overall path of the path component where the error occurred is
incorrect. More specifically, if the windows path being parsed contains
a prefix that includes a backslash or foreslash character (which is the
majority of prefixes), then the position of that first
backslash/foreslash is stored as the component end position in the
error regardless of whether the prefix is the component that caused the
error or not.

This change fixes the issue ensuring that the end position stored in the
error is equal to or greater than the position that the parser error
occurred at.