fix: errors parsing windows path prefix being ignored
With only the parser-iter feature active, given a windows path that
includes a prefix component, when iterating over the components of a
windows path, if an error occurs parsing the path's prefix, then the
prefix is interpreted as a normal path component. This is incorrect and
leads to inaccurate reporting of parsing errors.

The cause of this issue is that parsing a windows path prefix does not
detect errors that occur.

This change fixes the issue by explicitly handling any parsing errors
that occur when processing a windows path's prefix.