No longer store start and end positions in error objects
Originally, it was intended that the start and end positions of a
component in the path are stored in the error objects. The idea was to
be able to display the path component where the error occurred. However,
this turned out to be unnecessary once the error position (ie the
position in the path where a parsing error occurred) was added to error
objects; the path and the position of an error in the path are all that
is needed in order to tell a user where the issue is in the path. On top
of that, different implementations of path iteration had differing
capabilities wrt reporting positions; some implementations could
reliably report the exact start, end, and error positions of a path
component that contained a parsing error, while other implementations
could only reliably record the start and error positions.

This change no longer stored start and end positions in error objects.
Now, error objects only store the path, an error message fragment, and
the position in the path where the error occurred.